Remove deprecated StandardOutput=syslog
This is now equivalent to StandardOutput=journal, and recent versions of systemd log a warning for it. Instead of updating it to journal, remove it: the default if not specified is to respect the DefaultStandardOutput setting from /etc/systemd/system.conf, which in turn defaults to journal. Removing the explicit setting here means gdm is included if a sysadmin edits system.conf to change the logging destination for (almost) all system services at the same time, which seems desirable.

Similarly, StandardError defaults to DefaultStandardError, which in turn defaults to inherit, so we don't need to override it here.